Pull rel link methods into a trait
