+++ src/Entity/Collection/EntityCollection.php
@@ -8,6 +8,7 @@ use Symfony\Component\Serializer\Encoder\JsonEncoder;
 
 abstract class EntityCollection implements \Iterator, \Countable
 {
+    use CollectionRelLinkTrait;
 
     private $iteratorPosition = 0;
 
@@ -33,51 +34,6 @@ abstract class EntityCollection implements \Iterator, \Countable
     public static function fromResponse(ResponseInterface $response)
     {
         return new static((new JsonDecode())->decode((string) $response->getBody(), JsonEncoder::FORMAT));
-    }
-
-    public function getSelfLink()
-    {
-        return $this->getRelLink('self');
-    }
-
-    public function getFirstLink()
-    {
-        return $this->getRelLink('first');
-    }
-
-    public function getPreviousLink()
-    {
-        return $this->getRelLink('prev');
-    }
-
-    public function getNextLink()
-    {
-        return $this->getRelLink('next');
-    }
-
-    public function getLastLink()
-    {
-        return $this->getRelLink('last');
-    }
-
-    /**
-     * Get the related link.
-     *
-     * @param string $link_name
-     *   The name of the related link to retrieve.
-     *
-     * @return bool|\Psr\Http\Message\UriInterface
-     *   The related URL or FALSE if not present.
-     */
-    protected function getRelLink($link_name)
-    {
-        if (!isset($this->rawData->$link_name)) {
-            return false;
-        }
-
-        $uri = new Uri($this->rawData->$link_name);
-        $uri = $uri->withPath($uri->getPath() . '.json');
-        return $uri;
     }
 
     /**
